Lets use the proper term legal notice
From wikipedia https://en.wikipedia.org/wiki/Impressum:
The terms "masthead" and "colophon" apply to printed publications only
and are not commonly used on English-language websites, while "site
notice" is website-specific and "legal notice" or "legal disclosure" are
rarely found in printed works. An "imprint" in publishing may also mean
a brand name under which a work is published,[6] and so may not be
understood to mean an Impressum.

Signed-off-by: Artem Sidorenko <artem@posteo.de>